incubyte-backend-coding-exercise-main/meta/requirements.txt
click>=8.1.7,<9.0
PyGithub>=2.1.1,<3.0 # used for github API - repo creation, invites
GitPython>=3.1.37,<4.0 # used for git operations - clone, push